+++ src/ImageData/FileLoader.h
@@ -213,6 +213,9 @@ public:
     virtual void SetStokesCdelt(int stokes_cdelt);
     virtual bool GetStokesTypeIndex(const CARTA::PolarizationType& stokes_type, int& stokes_index);
 
+    // Modify time changed
+    bool ImageUpdated();
+
 protected:
     // Full name and hdu of the image file
     std::string _filename;
@@ -259,9 +262,6 @@ protected:
 
     // Basic flux density calculation
     double CalculateBeamArea();
-
-    // Modify time changed
-    bool ImageUpdated();
 };
 
 } // namespace carta