+++ test/TestHdf5Image.cc
@@ -17,7 +17,7 @@ using namespace carta;
 // Allows testing of protected methods in Frame without polluting the original class
 class TestFrame : public Frame {
 public:
-    TestFrame(uint32_t session_id, carta::FileLoader* loader, const std::string& hdu, int default_z = DEFAULT_Z)
+    TestFrame(uint32_t session_id, std::shared_ptr<carta::FileLoader> loader, const std::string& hdu, int default_z = DEFAULT_Z)
         : Frame(session_id, loader, hdu, default_z) {}
     FRIEND_TEST(Hdf5ImageTest, ExampleFriendTest);
 };
@@ -26,9 +26,9 @@ class Hdf5ImageTest : public ::testing::Test, public ImageGenerator {};
 
 TEST_F(Hdf5ImageTest, BasicLoadingTest) {
     auto path_string = GeneratedHdf5ImagePath("10 10");
-    std::unique_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(path_string));
+    std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(path_string));
     EXPECT_NE(loader.get(), nullptr);
-    std::unique_ptr<Frame> frame(new Frame(0, loader.release(), "0"));
+    std::unique_ptr<Frame> frame(new Frame(0, loader, "0"));
     EXPECT_NE(frame.get(), nullptr);
     EXPECT_TRUE(frame->IsValid());
 }
@@ -36,14 +36,16 @@ TEST_F(Hdf5ImageTest, BasicLoadingTest) {
 TEST_F(Hdf5ImageTest, ExampleFriendTest) {
     auto path_string = GeneratedHdf5ImagePath("10 10");
     // TestFrame used instead of Frame if access to protected values is required
-    std::unique_ptr<TestFrame> frame(new TestFrame(0, carta::FileLoader::GetLoader(path_string), "0"));
+    std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(path_string));
+    std::unique_ptr<TestFrame> frame(new TestFrame(0, loader, "0"));
     EXPECT_TRUE(frame->IsValid());
     EXPECT_TRUE(frame->_open_image_error.empty());
 }
 
 TEST_F(Hdf5ImageTest, CorrectShape2dImage) {
     auto path_string = GeneratedHdf5ImagePath("10 10");
-    std::unique_ptr<Frame> frame(new Frame(0, carta::FileLoader::GetLoader(path_string), "0"));
+    std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(path_string));
+    std::unique_ptr<Frame> frame(new Frame(0, loader, "0"));
     EXPECT_TRUE(frame->IsValid());
 
     auto shape = frame->ImageShape();
@@ -56,7 +58,8 @@ TEST_F(Hdf5ImageTest, CorrectShape2dImage) {
 
 TEST_F(Hdf5ImageTest, CorrectShape3dImage) {
     auto path_string = GeneratedHdf5ImagePath("10 10 10");
-    std::unique_ptr<Frame> frame(new Frame(0, carta::FileLoader::GetLoader(path_string), "0"));
+    std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(path_string));
+    std::unique_ptr<Frame> frame(new Frame(0, loader, "0"));
     EXPECT_TRUE(frame->IsValid());
 
     auto shape = frame->ImageShape();
@@ -71,7 +74,8 @@ TEST_F(Hdf5ImageTest, CorrectShape3dImage) {
 
 TEST_F(Hdf5ImageTest, CorrectShapeDegenerate3dImages) {
     auto path_string = GeneratedHdf5ImagePath("10 10 10 1");
-    std::unique_ptr<Frame> frame(new Frame(0, carta::FileLoader::GetLoader(path_string), "0"));
+    std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(path_string));
+    std::unique_ptr<Frame> frame(new Frame(0, loader, "0"));
     EXPECT_TRUE(frame->IsValid());
 
     auto shape = frame->ImageShape();
@@ -86,7 +90,8 @@ TEST_F(Hdf5ImageTest, CorrectShapeDegenerate3dImages) {
 
     // CASA-generated images often have spectral and Stokes axes swapped
     path_string = GeneratedHdf5ImagePath("10 10 1 10");
-    frame.reset(new Frame(0, carta::FileLoader::GetLoader(path_string), "0"));
+    loader.reset(carta::FileLoader::GetLoader(path_string));
+    frame.reset(new Frame(0, loader, "0"));
     EXPECT_TRUE(frame->IsValid());
 
     shape = frame->ImageShape();
@@ -102,7 +107,8 @@ TEST_F(Hdf5ImageTest, CorrectShapeDegenerate3dImages) {
 
 TEST_F(Hdf5ImageTest, CorrectShape4dImages) {
     auto path_string = GeneratedHdf5ImagePath("10 10 5 2");
-    std::unique_ptr<Frame> frame(new Frame(0, carta::FileLoader::GetLoader(path_string), "0"));
+    std::shared_ptr<carta::FileLoader> loader(carta::FileLoader::GetLoader(path_string));
+    std::unique_ptr<Frame> frame(new Frame(0, loader, "0"));
     EXPECT_TRUE(frame->IsValid());
 
     auto shape = frame->ImageShape();
@@ -117,7 +123,8 @@ TEST_F(Hdf5ImageTest, CorrectShape4dImages) {
 
     // CASA-generated images often have spectral and Stokes axes swapped
     path_string = GeneratedHdf5ImagePath("10 10 2 5");
-    frame.reset(new Frame(0, carta::FileLoader::GetLoader(path_string), "0"));
+    loader.reset(carta::FileLoader::GetLoader(path_string));
+    frame.reset(new Frame(0, loader, "0"));
     EXPECT_TRUE(frame->IsValid());
 
     shape = frame->ImageShape();